+++ src/test/java/lotto/domain/MoneyTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
 
-import static lotto.domain.LottoNumber.GAME_PRICE;
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
